+++ android/src/main/java/io/parity/signer/screens/keysets/create/NewSeedMenu.kt
@@ -1,9 +1,7 @@
-package io.parity.signer.screens.keysets
+package io.parity.signer.screens.keysets.create
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
 import androidx.compose.ui.Modifier
