+++ android/src/main/java/io/parity/signer/ui/navigationselectors/ScreenSelectorsNew.kt
@@ -10,7 +10,7 @@ import io.parity.signer.models.*
 import io.parity.signer.screens.keydetails.KeyDetailsMenuAction
 import io.parity.signer.screens.keydetails.exportprivatekey.PrivateKeyExportBottomSheet
 import io.parity.signer.screens.keydetails.KeyDetailsPublicKeyScreen
-import io.parity.signer.screens.keysets.NewSeedMenu
+import io.parity.signer.screens.keysets.create.NewSeedMenu
 import io.parity.signer.screens.logs.LogsMenu
 import io.parity.signer.screens.logs.LogsScreen
 import io.parity.signer.screens.logs.toLogsScreenModel
